+++ src/vue/forms/LimitsForm.vue
@@ -29,23 +29,42 @@
         <tbody>
           <tr>
             <td>{{ 'limits-form.daily-limit-lbl' | globalize }}</td>
-            <!-- eslint-disable-next-line -->
-            <td>{{ selectedLimitsByOpType.dailyOut || 'limits-form.not-set-lbl' | globalize }}</td>
+            <td>
+              {{
+                getLimitByScope('dailyOut') ||
+                  'limits-form.not-set-lbl' | globalize
+              }}
+            </td>
           </tr>
           <tr>
             <td>{{ 'limits-form.weekly-limit-lbl' | globalize }}</td>
             <!-- eslint-disable-next-line -->
-            <td>{{ selectedLimitsByOpType.weeklyOut || 'limits-form.not-set-lbl' | globalize }}</td>
+            <td>
+              {{
+                getLimitByScope('weeklyOut') ||
+                  'limits-form.not-set-lbl' | globalize
+              }}
+            </td>
           </tr>
           <tr>
             <td>{{ 'limits-form.monthly-limit-lbl' | globalize }}</td>
             <!-- eslint-disable-next-line -->
-            <td>{{ selectedLimitsByOpType.monthlyOut || 'limits-form.not-set-lbl' | globalize }}</td>
+            <td>
+              {{
+                getLimitByScope('monthlyOut') ||
+                  'limits-form.not-set-lbl' | globalize
+              }}
+            </td>
           </tr>
           <tr>
             <td>{{ 'limits-form.annual-limit-lbl' | globalize }}</td>
             <!-- eslint-disable-next-line -->
-            <td>{{ selectedLimitsByOpType.annualOut || 'limits-form.not-set-lbl' | globalize }}</td>
+            <td>
+              {{
+                getLimitByScope('annualOut') ||
+                  'limits-form.not-set-lbl' | globalize
+              }}
+            </td>
           </tr>
         </tbody>
       </table>
@@ -209,12 +228,17 @@ const EVENTS = {
 }
 
 const MIN_VALID_LIMIT_VALUE = 0
+const MAX_VALID_LIMIT_VALUE = config.MAX_AMOUNT
 
 export default {
   name: 'limits-form',
   mixins: [FormMixin],
   props: {
-    limits: { type: Object, required: true, default: () => ({}) },
+    limits: {
+      type: Object,
+      required: true,
+      default: () => ({}),
+    },
   },
   data: () => ({
     form: {
@@ -313,6 +337,31 @@ export default {
       if (!this.isFormValid()) return
       this.showConfirmation()
     },
+    getLimitByScope (period) {
+      const limitByType = this.selectedLimitsByOpType
+      switch (period) {
+        case 'dailyOut':
+          if (limitByType.dailyOut === MAX_VALID_LIMIT_VALUE) {
+            return ''
+          }
+          return limitByType.dailyOut
+        case 'weeklyOut':
+          if (limitByType.weeklyOut === MAX_VALID_LIMIT_VALUE) {
+            return ''
+          }
+          return limitByType.weeklyOut
+        case 'monthlyOut':
+          if (limitByType.monthlyOut === MAX_VALID_LIMIT_VALUE) {
+            return ''
+          }
+          return limitByType.monthlyOut
+        case 'annualOut':
+          if (limitByType.annualOut === MAX_VALID_LIMIT_VALUE) {
+            return ''
+          }
+          return limitByType.annualOut
+      }
+    },
     async submit () {
       this.disableForm()
       this.isRequestCreating = true
@@ -372,7 +421,7 @@ export default {
 
 .limits-form__form-subheading {
   margin-top: 3.2rem;
-  margin-bottom: -.8rem;
+  margin-bottom: -0.8rem;
 }
 
 .limits__assets-select {