+++ src/vue/forms/TransferForm.vue
@@ -352,7 +352,7 @@ export default {
 
 <style lang="scss" scoped>
 @import './app-form';
-@import "~@scss/variables";
+@import '~@scss/variables';
 
 .transfer__fee-box {
   padding-top: 4rem;
@@ -361,7 +361,7 @@ export default {
 .transfer__fee-box-heading {
   margin-top: 0;
   margin-bottom: 1rem;
-  font-weight: normal;
+  font-weight: 400;
   display: block;
   font-size: 1.6rem;
   color: $col-text-page-heading;