+++ src/vue/navigation/Passport.vue
@@ -92,6 +92,8 @@ export default {
       kycAvatarKey: vuexTypes.kycAvatarKey,
 
       isAccountUnverified: vuexTypes.isAccountUnverified,
+      isAccountUsAccredited: vuexTypes.isAccountUsAccredited,
+      isAccountUsVerified: vuexTypes.isAccountUsVerified,
       isAccountCorporate: vuexTypes.isAccountCorporate,
       isAccountGeneral: vuexTypes.isAccountGeneral,
       isAccountBlocked: vuexTypes.isAccountBlocked,
@@ -103,6 +105,10 @@ export default {
         return 'passport.account-general'
       } else if (this.isAccountCorporate) {
         return 'passport.account-corporate'
+      } else if (this.isAccountUsAccredited) {
+        return 'passport.account-us-accredited'
+      } else if (this.isAccountUsVerified) {
+        return 'passport.account-us-verified'
       } else if (this.isAccountBlocked) {
         return 'passport.account-blocked'
       } else {
@@ -184,11 +190,11 @@ export default {
 }
 </script>
 
-<style scoped lang="scss">
-@import "~@scss/mixins";
-@import "~@scss/variables";
+<style lang="scss">
+@import '~@scss/variables';
+@import '~@scss/mixins';
 
-$hide-account-details-bp: 800px;
+$media-hide-account-details-bp: 800px;
 $dropdown-item-side-padding: 2.4rem;
 
 .passport {
@@ -223,7 +229,8 @@ $dropdown-item-side-padding: 2.4rem;
 
 .passport__account-details-wrp {
   margin-left: 1.6rem;
-  @include respond-to-custom($hide-account-details-bp) {
+
+  @include respond-to-custom($media-hide-account-details-bp) {
     display: none;
   }
 }
@@ -254,7 +261,7 @@ $dropdown-item-side-padding: 2.4rem;
 
 .passport__account-type--blocked {
   color: $col-error;
-  font-weight: bold;
+  font-weight: 700;
 }
 
 .passport__dropdown {
@@ -264,40 +271,42 @@ $dropdown-item-side-padding: 2.4rem;
   background: $col-block-bg;
   display: flex;
   flex-direction: column;
+
   @include box-shadow();
 
   &:before {
     // dropdown arrow
-    content: "";
+    content: '';
     position: absolute;
     width: 0;
     height: 0;
     border-style: solid;
-    border-width: 0 0.8rem 0.8rem 0.8rem;
+    border-width: 0 0.8rem 0.8rem;
     border-color: transparent transparent $col-block-bg transparent;
     top: -0.8rem;
     right: 3.9rem;
   }
+}
 
-  &-enter-active {
-    animation-duration: 0.2s;
-    animation: passport-dropdown-keyframes 0.2s ease-in-out;
+.passport__dropdown-enter-active {
+  animation: passport-dropdown-keyframes 0.2s ease-in-out;
+  animation-duration: 0.2s;
+}
+
+.passport__dropdown-leave-active {
+  animation: passport-dropdown-keyframes 0.2s ease-in-out reverse;
+  animation-duration: 0.2s;
+}
+
+@keyframes passport-dropdown-keyframes {
+  from {
+    opacity: 0;
+    margin-top: -1.2rem;
   }
 
-  &-leave-active {
-    animation-duration: 0.2s;
-    animation: passport-dropdown-keyframes 0.2s ease-in-out reverse;
-  }
-
-  @keyframes passport-dropdown-keyframes {
-    from {
-      opacity: 0;
-      margin-top: -1.2rem;
-    }
-    to {
-      opacity: 1;
-      margin-top: 0;
-    }
+  to {
+    opacity: 1;
+    margin-top: 0;
   }
 }
 
@@ -310,7 +319,7 @@ $dropdown-item-side-padding: 2.4rem;
 }
 
 .passport__dropdown-signed-in-email {
-  font-weight: bold;
+  font-weight: 700;
 }
 
 .passport__dropdown-status-icon {