+++ src/vue/pages/Fees.vue
@@ -133,7 +133,7 @@ export default {
 </script>
 
 <style lang="scss" scoped>
-@import "~@scss/variables";
+@import '~@scss/variables';
 
 .fees-renderer {
   width: 100%;