+++ src/vuex/kyc.module.js
@@ -113,8 +113,8 @@ export const actions = {
     commit(vuexTypes.SET_KYC_RELATED_REQUEST, request)
   },
 
-  async [vuexTypes.LOAD_KYC_DATA] ({ state, commit }) {
-    const latestBlobId = state.request.blobId || state.relatedRequest.blobId
+  async [vuexTypes.LOAD_KYC_DATA] ({ state, getters, commit }) {
+    const latestBlobId = getters[vuexTypes.kycLatestBlobId]
     if (!latestBlobId) {
       return
     }
@@ -132,6 +132,7 @@ export const getters = {
   [vuexTypes.kycRequestRejectReason]: state => state.request.rejectReason,
   [vuexTypes.kycRequestResetReason]: state => state.request.resetReason,
   [vuexTypes.kycRequestBlockReason]: state => state.request.blockReason,
+  [vuexTypes.kycRequestExternalDetails]: state => state.request.externalDetails,
 
   [vuexTypes.isAccountRoleReseted]: state => state.isAccountRoleReseted,
   [vuexTypes.kycAccountRoleToSet]: state => state.isAccountRoleReseted
@@ -142,6 +143,8 @@ export const getters = {
   },
 
   [vuexTypes.kycRequestId]: state => state.request.id,
+  [vuexTypes.kycLatestBlobId]:
+      state => state.request.blobId || state.relatedRequest.blobId,
   [vuexTypes.kycLatestData]: state => JSON.parse(state.latestData),
   [vuexTypes.kycAvatarKey]: state => safeGet(
     JSON.parse(state.latestData),